requirements.txt
sacrebleu
pytorch-nlp
torch
nltk
numpy